Fixed importing a component from mui using the full path. Made the creation of an empty case to display detailed information in a separate function.
