+++ tms-ts/src/components/testcases/table.suites.component.tsx
@@ -1,16 +1,15 @@
-import {
-    Table,
-    TableCell,
-    Collapse,
-    IconButton,
-    Checkbox,
-    Link,
-    Box
-} from "@mui/material";
+import Table from "@mui/material/Table";
+import TableCell from "@mui/material/TableCell";
+import Collapse from "@mui/material/Collapse";
+import IconButton from "@mui/material/IconButton";
+import Checkbox from "@mui/material/Checkbox";
+import Link from "@mui/material/Link";
+import Box from "@mui/material/Box";
 import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
 import React, {useEffect, useMemo, useState} from "react";
 import useStyles from "../../styles/styles";
-import {myCase, suite, treeSuite} from "./suites.component";
+import {suite, treeSuite} from "./suites.component";
+import {myCase} from "../models.interfaces";
 import KeyboardArrowRightIcon from '@mui/icons-material/KeyboardArrowRight';
 import DetailedCaseInfo from "./detailed.case.info.component";
 import EditIcon from '@mui/icons-material/Edit';
@@ -413,18 +412,7 @@ const TableSuites = (props: {
         if (detailedCaseInfo.show) {
             if (shownCase.show && detailedCaseInfo.myCase.id === shownCase.myCaseId && lastEditCase !== detailedCaseInfo.myCase.id) {
                 document.getElementById(shownCase.myCaseId + "Arrow")!.style.transform = ""
-                setDetailedCaseInfo({
-                    show: false, myCase: {
-                        id: -1,
-                        name: "",
-                        suite: -1,
-                        scenario: "",
-                        project: -1,
-                        setup: "",
-                        teardown: "",
-                        estimate: -1
-                    }
-                })
+                setDetailedCaseInfo(SuiteCaseService.getEmptyDetailedCaseInfo())
                 setShownCase({show: false, myCaseId: -1})
             } else if (lastEditCase !== detailedCaseInfo.myCase.id) {
                 document.getElementById(detailedCaseInfo.myCase.id + "Arrow")!.style.transform = 'rotate(180deg)'
@@ -437,18 +425,7 @@ const TableSuites = (props: {
             }
         } else if (shownCase.myCaseId >= 0) {
             document.getElementById(shownCase.myCaseId + "Arrow")!.style.transform = ""
-            setDetailedCaseInfo({
-                show: false, myCase: {
-                    id: -1,
-                    name: "",
-                    suite: -1,
-                    scenario: "",
-                    project: -1,
-                    setup: "",
-                    teardown: "",
-                    estimate: -1
-                }
-            })
+            setDetailedCaseInfo(SuiteCaseService.getEmptyDetailedCaseInfo())
             setShownCase({show: false, myCaseId: -1})
         }
     }, [detailedCaseInfo])
